application-backend-server/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
PyJWT[crypto]
requests